Trigger a manual check from the GUI
Mostly because it's fun to play with channels.

This is a bit crappy, since the manual triggering is a no-op in the 0-59
seconds after a timeout (which defaults to happen once every 10
minutes).

A better behavior would be to set a timer to go off at next tick, or in
one minute, whichever comes first. I'm not up for that right now.