Points for when there are requested changes
We probably don't want to spend time on others' PRs when they already
have (blocking) feedback. I don't think it matters if it's us or not
that is blocking the PR.

When someone else blocks our PR though, we should probably prioritize
responding to that.

The TODO about finding our own comments that someone else responded to,
will be important to award points to. I think even if the PR is blocked
(-10 in this patch), someone else wanting us to comment on the PR should
bump it up high, at least over 0 so that the default minPoints=0 (in the
i3blocks script) will pick it up.