Remove points for outgoing comments (#29)
This requires deleting the sqlite db, because we don't do migrations :(

The current point adjustment is -10 points but I think it must be
weighed against some of the other criteria. For now, I'm lucky that the
only outgoing comment I have is on a small (+10 points) PR, negating
that, making it land at 0 and becoming non-actionable. TBC.