feat: Render all negative PRs in grey, not red
Before this commit:
- green: 0 or more points
- red: -1 or less points, not buried
- grey: buried

After:
- green: 0 or more points
- grey: -1 or less points

There was a bit too many colors with green, red, grey (and yellow for
bury-action, light-grey background, dark-grey font). Now we're in
harmony.

There's no real difference in actions to take on a negative PR, vs one
that's buried - the buried ones are always in the bottom, since we sort
by points, so they will be out of the way.